Working on updating tests
